refactor(PrintRaceResult): modify methods to return result and print String
